Fix hang where skynet attack list was not reset
Previously the game would hang randomly, whether
many or few moves had been made. This was because
the AI attack list was not cleared between games,
and was out of sync with the gameboard.

The AI attack method would check against the
gameboard to see if any moves were left to make,
but then only make the move if it wasn't already
in the AI attack list. If the attack list was
already full, it would keep trying, hence the hang.